+++ stack--2022/2-foundation/error-utils/src/fields.ts
@@ -3,8 +3,9 @@ import { XError, XXError } from './types'
 export const STRICT_STANDARD_ERROR_FIELDS = new Set<keyof XError>([
 	// standard fields
 	// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/prototype
-	'name',
-	'message',
+	'name', // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/name
+	'message', // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/message
+	'cause' // ES2022 https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/cause
 ])
 
 export const QUASI_STANDARD_ERROR_FIELDS = new Set<keyof XError>([
@@ -14,7 +15,7 @@ export const QUASI_STANDARD_ERROR_FIELDS = new Set<keyof XError>([
 
 	// quasi-standard: followed by all browsers + node
 	// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/prototype
-	'stack',
+	'stack', // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/Stack
 ])
 
 export const COMMON_ERROR_FIELDS = new Set<keyof XError>([
@@ -25,9 +26,6 @@ export const COMMON_ERROR_FIELDS = new Set<keyof XError>([
 	// standard in node only:
 	'code', // https://medium.com/the-node-js-collection/node-js-errors-changes-you-need-to-know-about-dc8c82417f65
 	        // https://nodejs.org/dist/latest/docs/api/errors.html#errors_node_js_error_codes
-
-	// Chrome only https://chromestatus.com/feature/5727099325251584
-	'cause',
 
 	// non standard but widely used:
 	'statusCode', // express, https://gist.github.com/zcaceres/2854ef613751563a3b506fabce4501fd